dot_gifs/requirements.txt
blinker==1.6.2
click==8.1.5
Flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyYAML==6.0
Werkzeug==2.3.6
zipp==3.16.2
